+++ frontend/src/hooks/useDatePicker/useDatePickerSelect.ts
@@ -2,10 +2,9 @@
 import { formatDateToWeekDates, getDateParts } from '@/utils/date';
 import { getDaysInMonth } from '@/utils/date/calendar';
 
-import { useHighlightRange } from './useHighlightRange';
-
 interface UseDatePickerSelectProps {
   baseDate?: Date;
+  selectedDate: Date | null;
   handleDateSelect: (date: Date) => void;
   gotoPrevMonth: () => void;
   gotoNextMonth: () => void;
@@ -13,12 +12,11 @@ interface UseDatePickerSelectProps {
 
 export const useDatePickerSelect = ({
   baseDate,
+  selectedDate,
   handleDateSelect,
   gotoPrevMonth,
   gotoNextMonth,
 }: UseDatePickerSelectProps) => {
-  const { highlightRange, setHighlightRange } = useHighlightRange();
-
   const onDateCellClick = (date: Date) => {
     const { year, month } = getDateParts(baseDate!);
     const daysInMonth = getDaysInMonth(baseDate!);
@@ -32,10 +30,15 @@ export const useDatePickerSelect = ({
     }
     
     handleDateSelect(date);
-    const weekDates = formatDateToWeekDates(date);
-    const [startOfWeek, endOfWeek] = [weekDates[0], weekDates[weekDates.length - 1]];
-    setHighlightRange({ start: startOfWeek, end: endOfWeek });
   };
 
-  return { highlightRange, onDateCellClick };
+  const getHighlightRange = () => {
+    const weekDates = formatDateToWeekDates(selectedDate);
+    if (!weekDates) return { start: null, end: null };
+    
+    const [startOfWeek, endOfWeek] = [weekDates[0], weekDates[weekDates.length - 1]];
+    return { start: startOfWeek, end: endOfWeek };
+  };
+
+  return { highlightRange: getHighlightRange(), onDateCellClick };
 };
